UPSTREAM: Remove --enable-webgl-software-rendering
Most probably it is useless with the new default ANGLE backend.
The original code has been trimmed and now it is just equivalent with
passing the following command line arguments:
--webEngineArgs --disable-gpu-rasterization --ignore-gpu-blocklist

As a drive-by, change WebEngineContext::initCommandLine() parameter to
pointer instead of reference to make it clear on the caller side it will
not use but it is expected to change the parameter's value.

Pick-to: 6.10
Change-Id: Ic118fa55071c5d8ac0730db2f640c3132375db8b
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>